refactored logging and config locations
