+++ CONTRIBUTORS.txt
@@ -18,11 +18,6 @@
     and the tests as a part of his work on the Open Tree of Life
     api.opentreeoflife.org repository.
 
-    The gcmdr functionality in peyotl is a direct port of the gcmdr
-    from https://github.com/OpenTreeOfLife/gcmdr with lots of code
-    copyied over. That code was written by Stephen A. Smith 
-    and Joseph W. Brown
-
 ----------
 
 N: Emily Jane McTavish
