+++ doc/peyotl_writeup.tex
@@ -133,9 +133,8 @@ of an Open Tree of Life API adaptors \citep{sharedapitests}.
     tests written by that group is executed as a part of \pey's regular regression testing.
 
 
-Some nascent subpackages (`ott' and `gcmdr') of \pey are in development as re-implementations
-    of other \otol tools (for taxonomic operations and driving of the ``treemachine'' supertree tool
-    respectively).
+Some nascent subpackages (`ott') of \pey are in development as re-implementations
+    of other \otol tools (for taxonomic operations).
 These parts of \pey will provide a means of testing the canonical implementations of these
     tools, but will also offer more flexible interfaces for python programmers.
