--- extras/gcmdr/asterales/config
@@ -1,26 +1,0 @@
-[treemachine]
-# filepath for taxonomy database
-tax_db = scratch/gcmdr/asterales/taxo.db
-
-# start of java invocation
-java = java -Xmx8g
-
-# ott id for starting synthesis
-synth_ott_id = 1042120
-
-# filepath for loading studies
-studies_db= scratch/gcmdr/asterales/asterales_loading.db
-
-# treemachine log file
-log = scratch/gcmdr/asterales/logfile
-
-# filepath for tree
-tree_file = scratch/gcmdr/asterales/asterales_synth.tre
-
-# nexson cache
-nexson_cache = /tmp/gcmdr_nexsons
-
-# tree lists
-tree_list_files = extras/gcmdr/asterales/studies.txt
-
-
