--- extras/gcmdr/asterales/gcmdr-repo.config
@@ -1,35 +1,0 @@
-# this file is intended to be run from the top
-# of the gcmdr repo (but with using the peyotl impl of 
-# gcmdr
-[treemachine]
-# filepath for taxonomy database
-tax_db = example/asterales_taxonomy.db
-# filepath for loading studies
-load_db = example/asterales_loading.db
-# filepath for synthesis
-synthesis_db = example/asterales_synth.db
-
-# filepath for storing trees that were loaded
-loaded_trees_json = example/loaded_trees.json
-
-# start of java invocation
-java = java -Xmx8g
-
-# ott id for starting synthesis
-synth_ott_id = 1042120
-
-
-# treemachine log file
-log = example/logfile
-
-# filepath for tree
-out_tree_file = example/asterales_synth.tre
-
-# nexson cache
-nexson_cache = example/nexsons
-
-# tree lists
-tree_list_files = extras/gcmdr/asterales/studies.txt
-
-[ott]
-parent = example/aster
