--- extras/gcmdr/asterales/studies.txt
@@ -1,9 +1,0 @@
-pg_2539_6294    # Soltis et al. 2011
-pg_715_1289     # Barnadesioideae
-pg_329_324      # Hieracium
-#pg_9_1         # Campanulidae
-pg_1944_3959    # Campanulidae; replacement of 9 above
-pg_709_1276     # Lobelioideae, very non monophyletic
-pg_41_1396      # Feddea
-pg_82_5792      # Campanula, very non monophyletic campanula
-pg_932_1831     # Goodeniaceae, tons of non monophyly
