--- scratch/gcmdr/README
@@ -1,1 +1,0 @@
-This is the default scratch directory for gcmdr operations
